+++ etc/blocks/monetization-test/block/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @milaboratories/milaboratories.monetization-test
 
+## 1.0.130
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.19
+
+## 1.0.129
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.18
+
+## 1.0.128
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.17
+- @milaboratories/milaboratories.monetization-test.model@1.0.22
+- @milaboratories/milaboratories.monetization-test.ui@1.1.109
+
 ## 1.0.127
 
 ### Patch Changes