+++ etc/blocks/pool-explorer/workflow/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @milaboratories/milaboratories.pool-explorer.workflow
 
+## 1.1.73
+
+### Patch Changes
+
+- Updated dependencies [60083a1]
+  - @platforma-sdk/workflow-tengo@5.3.6
+
+## 1.1.72
+
+### Patch Changes
+
+- Updated dependencies [15d75b4]
+  - @platforma-sdk/workflow-tengo@5.3.5
+
+## 1.1.71
+
+### Patch Changes
+
+- Updated dependencies [bb07805]
+  - @platforma-sdk/workflow-tengo@5.3.4
+
 ## 1.1.70
 
 ### Patch Changes