+++ etc/blocks/ui-examples/block/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.399
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.73
+
+## 1.0.398
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.72
+
+## 1.0.397
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.71
+- @platforma-sdk/model@1.44.8
+- @milaboratories/milaboratories.ui-examples.model@1.2.44
+- @milaboratories/milaboratories.ui-examples.ui@1.5.50
+
 ## 1.0.396
 
 ### Patch Changes