+++ etc/blocks/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.5.50
+
+### Patch Changes
+
+- @platforma-sdk/model@1.44.8
+- @milaboratories/milaboratories.ui-examples.model@1.2.44
+
 ## 1.5.49
 
 ### Patch Changes