+++ lib/model/backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/pl-model-backend
 
+## 1.1.17
+
+### Patch Changes
+
+- Updated dependencies [7af7faf]
+  - @milaboratories/pl-client@2.16.0
+
+## 1.1.16
+
+### Patch Changes
+
+- @milaboratories/pl-client@2.15.1
+
 ## 1.1.15
 
 ### Patch Changes