+++ lib/node/pl-errors/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/pl-errors
 
+## 1.1.32
+
+### Patch Changes
+
+- Updated dependencies [7af7faf]
+  - @milaboratories/pl-client@2.16.0
+
+## 1.1.31
+
+### Patch Changes
+
+- @milaboratories/pl-client@2.15.1
+
 ## 1.1.30
 
 ### Patch Changes