+++ sdk/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @platforma-sdk/workflow-tengo
 
+## 5.3.6
+
+### Patch Changes
+
+- 60083a1: Parquet write
+- Updated dependencies [60083a1]
+  - @platforma-open/milaboratories.software-ptabler@1.12.7
+
+## 5.3.5
+
+### Patch Changes
+
+- 15d75b4: FlatBuffer validation options maxed
+- Updated dependencies [15d75b4]
+  - @platforma-open/milaboratories.software-ptabler@1.12.6
+
+## 5.3.4
+
+### Patch Changes
+
+- bb07805: xsv.exportFrame migration to ptabler
+- Updated dependencies [bb07805]
+  - @platforma-open/milaboratories.software-ptabler@1.12.5
+
 ## 5.3.3
 
 ### Patch Changes